Update SmokingHabit.swift
cigarettePrice을 Double으로 타입 변경

+++ Sennight/Sennight/Models/SmokingHabit.swift
@@ -10,7 +10,7 @@ import Foundation
 struct SmokingHabitsRequest: Codable {
     let userID: Int
     let dailyCigarettes: Int
-    let cigarettePrice: Int
+    let cigarettePrice: Double
     let firstCigarette: String
     let smokingYears: Int
